+++ spec/integration/image-streamer/api300/build_plan/update_spec.rb
@@ -11,22 +11,22 @@
 
 require 'spec_helper'
 
-klass = OneviewSDK::ImageStreamer::API300::PlanScript
+klass = OneviewSDK::ImageStreamer::API300::BuildPlan
 RSpec.describe klass, integration_i3s: true, type: UPDATE do
   include_context 'integration i3s api300 context'
 
   describe '#update' do
     it 'updates the name of the plan script' do
-      item = klass.find_by($client_i3s_300, name: PLAN_SCRIPT1_NAME).first
-      expect(item['uri']).to be
-      item['name'] = PLAN_SCRIPT1_NAME_UPDATE
+      item = klass.find_by($client_i3s_300, name: BUILD_PLAN1_NAME).first
+      expect(item).to be
+      item['name'] = BUILD_PLAN1_NAME_UPDATED
       expect { item.update }.not_to raise_error
       item.retrieve!
-      expect(item['name']).to eq(PLAN_SCRIPT1_NAME_UPDATE)
-      item['name'] = PLAN_SCRIPT1_NAME
+      expect(item['name']).to eq(BUILD_PLAN1_NAME_UPDATED)
+      item['name'] = BUILD_PLAN1_NAME
       expect { item.update }.not_to raise_error
       item.retrieve!
-      expect(item['name']).to eq(PLAN_SCRIPT1_NAME)
+      expect(item['name']).to eq(BUILD_PLAN1_NAME)
     end
   end
 end