+++ spec/integration/image-streamer/api300/golden_image/delete_spec.rb
@@ -11,22 +11,20 @@
 
 require 'spec_helper'
 
-klass = OneviewSDK::ImageStreamer::API300::PlanScript
-RSpec.describe klass, integration_i3s: true, type: UPDATE do
+klass = OneviewSDK::ImageStreamer::API300::GoldenImage
+RSpec.describe klass, integration_i3s: true, type: DELETE, sequence: i3s_rseq(klass) do
   include_context 'integration i3s api300 context'
 
-  describe '#update' do
-    it 'updates the name of the plan script' do
-      item = klass.find_by($client_i3s_300, name: PLAN_SCRIPT1_NAME).first
-      expect(item['uri']).to be
-      item['name'] = PLAN_SCRIPT1_NAME_UPDATE
-      expect { item.update }.not_to raise_error
-      item.retrieve!
-      expect(item['name']).to eq(PLAN_SCRIPT1_NAME_UPDATE)
-      item['name'] = PLAN_SCRIPT1_NAME
-      expect { item.update }.not_to raise_error
-      item.retrieve!
-      expect(item['name']).to eq(PLAN_SCRIPT1_NAME)
+  describe '#delete' do
+    it 'removes all golden images' do
+      item = klass.find_by($client_i3s_300, name: GOLDEN_IMAGE1_NAME).first
+      item2 = klass.find_by($client_i3s_300, name: GOLDEN_IMAGE2_NAME).first
+      expect(item).to be
+      expect(item2).to be
+      expect { item.delete }.not_to raise_error
+      expect(item.retrieve!).to eq(false)
+      expect { item2.delete }.not_to raise_error
+      expect(item2.retrieve!).to eq(false)
     end
   end
 end